Move model tests to their own folder.
