+++ run_tests.py
@@ -21,35 +21,35 @@ from tests.logic_layer.task_set_deleted_tests import *
 from tests.logic_layer.task_unset_done_tests import *
 from tests.logic_layer.task_unset_deleted_tests import *
 from tests.logic_layer.do_reset_order_nums_tests import *
-from tests.tag_tests import *
-from tests.task_tests import *
-from tests.task_id_tests import *
-from tests.task_css_tests import *
-from tests.task_tags_tests import *
-from tests.task_duration_text_tests import *
-from tests.task_cost_text_tests import *
+from tests.models_t.tag_tests import *
+from tests.models_t.task_tests import *
+from tests.models_t.task_id_tests import *
+from tests.models_t.task_css_tests import *
+from tests.models_t.task_tags_tests import *
+from tests.models_t.task_duration_text_tests import *
+from tests.models_t.task_cost_text_tests import *
 from tests.logic_layer.search_tests import *
-from tests.task_dependencies_tests import *
+from tests.models_t.task_dependencies_tests import *
 from tests.logic_layer.task_dependencies_tests import *
 from tests.logic_layer.task_tags_tests import *
 from tests.logic_layer.get_lowest_highest_order_num import *
-from tests.task_prioritize_tests import *
+from tests.models_t.task_prioritize_tests import *
 from tests.logic_layer.task_prioritize_tests import *
 from tests.view_layer.route_tests import *
 from tests.persistence_layer_tests import *
 from tests.logic_layer.do_import_data_tests import *
 from tests.logic_layer.set_task_tests import *
-from tests.task_interlinking_tests import *
-from tests.interlinking_tests import *
-from tests.task_from_dict_tests import *
-from tests.task_lazy_tests import *
-from tests.tag_from_dict_tests import *
-from tests.note_from_dict_tests import *
-from tests.note_tests import *
-from tests.attachment_tests import *
-from tests.attachment_from_dict_tests import *
-from tests.user_from_dict_tests import *
-from tests.option_from_dict_tests import *
+from tests.models_t.task_interlinking_tests import *
+from tests.models_t.interlinking_tests import *
+from tests.models_t.task_from_dict_tests import *
+from tests.models_t.task_lazy_tests import *
+from tests.models_t.tag_from_dict_tests import *
+from tests.models_t.note_from_dict_tests import *
+from tests.models_t.note_tests import *
+from tests.models_t.attachment_tests import *
+from tests.models_t.attachment_from_dict_tests import *
+from tests.models_t.user_from_dict_tests import *
+from tests.models_t.option_from_dict_tests import *
 
 
 def run():
